refactor(310p): scope KV cache budget scaling
Move multi-group KV cache budget scaling out of the generic NPUWorker and keep it on NPUWorker310, where the per-layer allocation requires it. Move the regression coverage to the 310P test suite.
